Retry executor test cleanup
