+++ golem-worker-executor/tests/lib.rs
@@ -18,7 +18,7 @@ use golem_worker_executor_test_utils::{
     test_component,
 };
 use std::fmt::Debug;
-use test_r::{sequential_suite, tag_suite, test_dep, timeout_suite};
+use test_r::{sequential_suite, tag_suite, test_dep};
 
 pub mod agent;
 pub mod agent_sdk_ts;
@@ -41,6 +41,7 @@ pub mod rdbms;
 pub mod rdbms_service;
 pub mod readonly;
 pub mod resource_limits;
+pub mod retry_lifecycle;
 pub mod retry_policies;
 pub mod revert;
 pub mod rpc;
@@ -79,6 +80,7 @@ pub fn raw_params(
 }
 
 tag_suite!(api, group1);
+tag_suite!(retry_lifecycle, group1);
 tag_suite!(blobstore, group1);
 tag_suite!(keyvalue, group1);
 tag_suite!(in_function_retry, group1);
@@ -109,8 +111,6 @@ tag_suite!(storage_quota, group1);
 sequential_suite!(key_value_storage);
 sequential_suite!(namespace_routed_key_value_storage);
 sequential_suite!(indexed_storage);
-
-timeout_suite!(in_function_retry, "2 minutes");
 
 #[derive(Debug)]
 pub struct Tracing;
