+++ golem-worker-executor/tests/retry_policies.rs
@@ -21,6 +21,7 @@ use golem_worker_executor_test_utils::{
     LastUniqueId, PrecompiledComponent, TestContext, WorkerExecutorTestDependencies, start,
 };
 use test_r::{inherit_test_dep, test, timeout};
+use std::time::{Duration, SystemTime};
 
 inherit_test_dep!(WorkerExecutorTestDependencies);
 inherit_test_dep!(LastUniqueId);
@@ -359,5 +360,70 @@ async fn list_retry_policy_names_returns_all(
     executor.check_oplog_is_queryable(&agent_id).await?;
 
     drop(executor);
+    Ok(())
+}
+
+#[test]
+#[tracing::instrument]
+async fn golem_rust_set_retry_policy(
+    last_unique_id: &LastUniqueId,
+    deps: &WorkerExecutorTestDependencies,
+    #[tagged_as("host_api_tests")] host_api_tests: &PrecompiledComponent,
+    _tracing: &Tracing,
+) -> anyhow::Result<()> {
+    let context = TestContext::new(last_unique_id);
+    let executor = start(deps, &context).await?;
+
+    let component = executor
+        .component_dep(&context.default_environment_id, host_api_tests)
+        .store()
+        .await?;
+
+    let agent_id = agent_id!("GolemHostApi", "set-retry-policy-1");
+    let worker_id = executor
+        .start_agent(&component.id, agent_id.clone())
+        .await?;
+
+    let mut _log_output_guards = Vec::new();
+    _log_output_guards.push(executor.log_output_scoped(&worker_id).await?);
+
+    let start = SystemTime::now();
+    let result1 = executor
+        .invoke_and_await_agent(
+            &component,
+            &agent_id,
+            "fail_with_custom_max_retries",
+            data_value!(2u64),
+        )
+        .await;
+    let elapsed = start.elapsed().unwrap();
+
+    let result2 = executor
+        .invoke_and_await_agent(
+            &component,
+            &agent_id,
+            "fail_with_custom_max_retries",
+            data_value!(1u64),
+        )
+        .await;
+
+    executor.check_oplog_is_queryable(&worker_id).await?;
+
+    assert!(elapsed < Duration::from_secs(3)); // 2 retry attempts, 1s delay
+    assert!(result1.is_err());
+    assert!(result2.is_err());
+    let result1_err = format!("{}", result1.unwrap_err());
+    assert!(
+        result1_err.contains("error while executing at wasm backtrace:")
+            || result1_err.contains("Invocation failed"),
+        "Unexpected error: {result1_err}"
+    );
+    let result2_err = format!("{}", result2.unwrap_err());
+    assert!(
+        result2_err.contains("Previous invocation failed")
+            || result2_err.contains("error while executing at wasm backtrace:"),
+        "Unexpected error: {result2_err}"
+    );
+
     Ok(())
 }
